Fix 4c5024627e: Apply padding to the "Next effect" text only
The way 4c5024627e added padding to the next effect text caused the income and effect displays in the hero and skill tables to become misaligned, as padding was added to not only the next effect text but also all other effect displays.

This is fixed by creating a new CSS class for just the next effect text and applying the padding there and there only.