Shorten exported saves by removing unnecessary data
In my tests, a save file that was 107 kB before this modification now exports as a 45.2-kB save, with no loss of functionality on the front end.

Adding new properties to `baseData` in the hero, skill, and item definitions will now no longer directly impact save file sizes.

`baseData` is still kept in the exported saves for compatibility, until a save version system is implemented.